refactor(middleware): remove frontegg's proxy middlewares
BREAKING CHANGE: support for outdated "frontegg"/"fronteggNextJs" middlewares is now dropped
