+++ .eslintrc.json
@@ -12,6 +12,12 @@
   "plugins": ["@typescript-eslint"],
   "rules": {
     "@typescript-eslint/no-explicit-any": ["off"],
+    "@typescript-eslint/no-empty-function": [
+      "error",
+      {
+        "allow": ["private-constructors"]
+      }
+    ],
     "max-len": [
       "error",
       {
