+++ src/clients/events/types/channel-configuration.ts
