+++ src/clients/events/types/errors.ts
