+++ src/clients/identity/index.ts
@@ -1,0 +1,2 @@
+export { IdentityClient } from './identity-client';
+export { withAuthentication } from '../../middleware/with-authentication';
