--- src/identity/index.ts
@@ -1,3 +1,0 @@
-export { IdentityClient } from './identity-client';
-export { withAuthentication } from './with-authentication';
-export { contextResolver, contextResolverWithPermissions } from './context-resolver';
