+++ src/middleware/index.ts
@@ -1,4 +1,1 @@
-export * from './frontegg-middleware';
-export * from './frontegg-nextjs';
-export * from './types';
-export * from './context-holder';
+export * from './with-authentication';
