+++ src/types/express/index.d.ts
@@ -1,0 +1,11 @@
+import { IUser } from '../../middleware';
+
+export {};
+
+declare global {
+  namespace Express {
+    export interface Request {
+      user?: IUser;
+    }
+  }
+}
