--- src/utils/sleep.ts
@@ -1,4 +1,0 @@
-export const sleep = (time: number) =>
-  new Promise((resolve) => {
-    setTimeout(resolve, time);
-  });
