+++ tsconfig.json
@@ -4,11 +4,13 @@
     "target": "es5",
     "module": "commonjs",
     "declaration": true,
+    "strictNullChecks": true,
+    "strictPropertyInitialization": true,
     "outDir": "./dist",
     "strict": true,
     "sourceMap": true,
     "skipLibCheck": true
   },
   "exclude": ["node_modules"],
-  "include": ["src"]
+  "include": ["src", "index.ts"]
 }
